Create a ProductList component and adjust its styling
