+++ src/components/ProductCard.tsx
@@ -12,15 +12,20 @@ function ProductCard({ product }: { product: Product }) {
     router.push(`/products/${product.id}`);
   };
   return (
-    <Card className="mb-[10px] product-card cursor-pointer m-[auto]" sx={{ maxWidth: 345 }} onClick={handleClick}>
+    <Card className="mb-[10px] product-card cursor-pointer m-[auto] relative" sx={{ maxWidth: 345 }} onClick={handleClick}>
+      <div className="product-discount">
+        <>{product.discountPercentage}%</>
+      </div>
       <CardMedia
         sx={{ height: 140 }}
-        image={product.images[0]}
-        title={product.images[0]}
+        image={product.thumbnail}
+        title={product.title}
       />
       <CardContent className="product-card-content">
+
         <h5 className="ellipse-text">{product.title}</h5>
         <p className="ellipse-2-lines">{product.description}</p>
+        <p className="text-[#EE4D2E] mt-3">฿{product.price}</p>
       </CardContent>
     </Card>
   )
